# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HMCA API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pms/rooms": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get all rooms for the property",
                "operationId": "27ec588efee8369f190d8811de74b45c",
                "responses": {
                    "200": {
                        "description": "List of rooms retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Create a new room",
                "operationId": "5687d5293545eb9ae33f2f6446019516",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "room_type_id",
                                    "room_number",
                                    "capacity"
                                ],
                                "properties": {
                                    "room_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "room_number": {
                                        "type": "string",
                                        "example": "A101"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "extra_capability": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "room_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2500.5
                                    },
                                    "bed_charge": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 300
                                    },
                                    "room_size": {
                                        "type": "string",
                                        "example": "single"
                                    },
                                    "bed_number": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "bed_type": {
                                        "type": "string",
                                        "example": "kingbed"
                                    },
                                    "room_description": {
                                        "type": "string",
                                        "example": "Deluxe AC Room"
                                    },
                                    "reserve_condition": {
                                        "type": "string",
                                        "example": "Non-refundable"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "amenity_ids": {
                                        "description": "Array of amenity IDs",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Room created successfully"
                    },
                    "400": {
                        "description": "Invalid data provided"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pms/rooms/{id}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get a room by ID",
                "operationId": "631a33972365a1273f8f3921db6db400",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room details fetched successfully"
                    },
                    "404": {
                        "description": "Room not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Update an existing room",
                "operationId": "b5fe77e0d75f05c0f3d436dc92ca11a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "room_type_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "room_number": {
                                        "type": "string",
                                        "example": "A202"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "room_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3500
                                    },
                                    "bed_type": {
                                        "type": "string",
                                        "example": "kingbed"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "extra_capability": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "bed_charge": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 400
                                    },
                                    "room_size": {
                                        "type": "string",
                                        "example": "single"
                                    },
                                    "bed_number": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "room_description": {
                                        "type": "string",
                                        "example": "Updated room description"
                                    },
                                    "reserve_condition": {
                                        "type": "string",
                                        "example": "Refundable"
                                    },
                                    "amenity_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Room updated successfully"
                    },
                    "404": {
                        "description": "Room not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Delete a room",
                "operationId": "664823375067c645a6bfe96be6c7f2bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room deleted successfully"
                    },
                    "404": {
                        "description": "Room not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}